pyproject.toml
[tool.poetry]
name = "reseller-cashback-api"
version = "0.1.0"
description = ""
authors = ["Giovani Moutinho <giovanim@ymail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
prettyconf = "^2.2.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
docker-compose = "^1.28.5"
autopep8 = "^1.5.6"
pytest = "^6.2.2"
pytest-django = "^4.1.0"
model-bakery = "^1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
